1Inch Solver (gnosis/gp-v2-services#520)
This PR adds an implementation for a 1Inch solver. It works by selecting a portion of orders and using the 1Inch API to directly get the transaction call data to use for an interaction to swap on 1Inch.

### Test Plan

Added a couple unit tests verifying the `solve` logic. Additionally one ignored test to get an actual settlement for a fictional order:
```
$ cargo test -p solver -- --ignored --nocapture solve_order_on_one
...
running 1 test
Settlement {
...
}
test oneinch_solver::tests::solve_order_on_oneinch ... ok
...
```

An end-to-end test that actually settles against 1Inch will be done when the solver is enabled.

+++ solver/src/liquidity/slippage.rs
@@ -3,10 +3,10 @@
 use ethcontract::U256;
 
 /// Constant maximum slippage of 10 BPS (0.1%) to use for on-chain liquidity.
-pub const MAX_SLIPPAGE_BPS: u32 = 10;
+pub const MAX_SLIPPAGE_BPS: u16 = 10;
 
 /// Basis points in 100%.
-const BPS_BASE: u32 = 10000;
+const BPS_BASE: u16 = 10000;
 
 /// Multiply an integer amount by a rational, with additional handling in case
 /// of overflows.
